[#11] refactor: 데이터 소스 분리
