+++ iOS/SideDish/SideDish/View/MainMenuViewDataSource.swift
@@ -1,0 +1,26 @@
+//
+//  MainMenuViewDataSource.swift
+//  SideDish
+//
+//  Created by 신한섭 on 2020/04/22.
+//  Copyright © 2020 신한섭. All rights reserved.
+//
+
+import UIKit
+
+class MainMenuViewDataSource: NSObject, UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return 10
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        guard let cell = tableView.dequeueReusableCell(withIdentifier:
+            "MainMenuCell", for: indexPath) as? MainMenuTableViewCell else {return UITableViewCell()}
+        cell.configuration()
+        return cell
+    }
+    
+    func numberOfSections(in tableView: UITableView) -> Int {
+        return 1
+    }
+}
